requirements.txt
opencv-python
pytesseract
pygetwindow
keyboard
pyautogui
pywin32
pillow